# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "Documentazione Swagger per la tua API Laravel",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/carts/{cart_id}": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Lista degli item del carrello",
                "description": "Restituisce una lista degli item all'interno del carrello",
                "operationId": "96deca4bfc699e15aa3a361a8424a48b",
                "parameters": [
                    {
                        "name": "cart_id",
                        "in": "path",
                        "description": "ID del carrello",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista degli item all'interno del carrello restituita con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista degli oggetti all'interno del carrello"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CartItem"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/carts/add": {
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "Aggiungi al cart un nuovo elemento",
                "description": "Aggiunge ad un cart (o lo crea) un singolo evento  ",
                "operationId": "24a3ebb22f948e8e4b6514f8e23ca61f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Oggetti creati con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Evento aggiunto al carrello"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CartItem"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/carts/remove": {
            "delete": {
                "tags": [
                    "Carts"
                ],
                "summary": "Rimuovi al cart un elemento",
                "description": "Rimuove ad un cart un singolo evento o una quantità prestabilita di biglietti  ",
                "operationId": "b798944d40d08c70c1bb01ededbcf3d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Oggetti Cancellati con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Evento aggiunto al carrello"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CartItem"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CartRequest": {
                "title": "Richiesta di Creazione o aggiunta di un item al carrello",
                "description": "Schema per la creazione o l'aggiunta di un prodotto al carrello",
                "required": [
                    "event_id, quantity"
                ],
                "properties": {
                    "cart_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "event_id": {
                        "type": "integer",
                        "example": "45"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": "25"
                    }
                },
                "type": "object"
            },
...
            "CartItem": {
                "title": "Item carrello",
                "description": "Schema di singolo item del carrello",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cart_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "event_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 1
                    },
                    "event": {
                        "$ref": "#/components/schemas/Event"
                    }
                },
                "type": "object"
            },
...
            "Event": {
                "title": "Evento",
                "description": "Schema di un evento",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Concerto Jazz"
                    },
                    "description": {
                        "type": "string",
                        "example": "Concerto jazz dal vivo"
                    },
                    "length": {
                        "type": "number",
                        "format": "float",
                        "example": 3
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-15 21:00:00"
                    },
                    "available_tickets": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    },
...
}